refactor: move Writables into their own file
